requirements.txt
# Cortex3d 依赖

# 图像处理
opencv-python>=4.8.0
Pillow>=10.0.0
rembg>=2.0.50
onnxruntime>=1.16.0

# Hugging Face API 调用
gradio_client>=0.10.0

# 3D 模型验证 (可选)
trimesh>=4.0.0
numpy>=1.24.0
